Move to resolvers/profiles
