+++ src/graphql/resolvers/repos.ts
@@ -1,6 +1,5 @@
 import { Arg, Ctx, Resolver, Query } from 'type-graphql';
-import { ClaimStatus, Profile, Repo } from '@generated/type-graphql';
-import { ProfileWithClaimsCount } from './profiles';
+import { Repo } from '@generated/type-graphql';
 import { getLastMonthStartDatetime } from './util';
 import { Context } from '../../context';
 import { createScopedLogger } from '../../logging';
@@ -75,48 +74,5 @@ export class CustomRepoResolver {
     endTimer({ success: 1 });
 
     return results;
-  }
-
-  @Query(returns => [ProfileWithClaimsCount])
-  async repoMostHonoredContributors(
-    @Ctx() { prisma }: Context,
-    @Arg('count', { defaultValue: 10 }) count: Number,
-    @Arg('repoId') repoId: number,
-  ): Promise<ProfileWithClaimsCount[]> {
-    const logger = createScopedLogger('GQL repoMostHonoredContributors');
-
-    logger.info(`Request for repo ${repoId}'s ${count} most honored contributors `);
-
-    const endTimer = gqlRequestDurationSeconds.startTimer('repoMostHonoredContributors');
-
-    type ResultType = Profile & {
-      claimsCount: Number;
-    };
-
-    const results: ResultType[] = await prisma.$queryRaw`
-      SELECT p.*, COUNT(c.id) AS "claimsCount"
-      FROM "Profile" AS p
-      JOIN "Claim" AS c ON c.address = p.address
-      JOIN "GitPOAP" AS gp ON gp.id = c."gitPOAPId"
-      WHERE c.status = ${ClaimStatus.CLAIMED}
-      AND gp."repoId" = ${repoId}
-      GROUP BY p.id
-      ORDER BY "claimsCount" DESC
-      LIMIT ${count}
-    `;
-
-    let finalResults = [];
-
-    for (const result of results) {
-      const { claimsCount, ...profile } = result;
-
-      finalResults.push({ profile, claimsCount });
-    }
-
-    logger.debug(`Completed request for repo ${repoId}'s ${count} most honored contributors`);
-
-    endTimer({ success: 1 });
-
-    return finalResults;
   }
 }
